Use branch name in version
When building something that is not the master (or main) brach, it is
important that the resulting artifacts should not be confused with
those from the master branch.

Given IMAGE_VERSION=47.11 and build number 17, a build from the master
branch should produce artifacts for version 47.11.17, but a build from
e.g. feature/new-stuff should produce artifacts for
47.11+feature.new.stuff.17.

The actual code change here is an untested suggestion for how to do this.